Refactor internal serializer construction / naming
