+++ src/ServiceControl.Plugin.Nsb6.SagaAudit/SagaAudit.cs
@@ -5,6 +5,7 @@
     using NServiceBus.Features;
     using NServiceBus.MessageInterfaces.MessageMapper.Reflection;
     using NServiceBus.Settings;
+    using NServiceBus.Transports;
     using Plugin;
     using Plugin.SagaAudit;
 
@@ -18,25 +19,15 @@
 
         protected override void Setup(FeatureConfigurationContext context)
         {
+            context.Container.ConfigureComponent<SagaAuditSerializer>(DependencyLifecycle.SingleInstance);
+
             context.Container.ConfigureComponent<ServiceControlBackend>(DependencyLifecycle.SingleInstance);
+            context.Container.ConfigureComponent<CaptureSagaStateBehavior>(DependencyLifecycle.SingleInstance);
 
             context.Pipeline.Register("CaptureSagaState", b => b.Build<CaptureSagaStateBehavior>(), "Records saga state changes");
             context.Pipeline.Register("ReportSagaStateChanges", new CaptureSagaResultingMessagesBehavior(), "Reports the saga state changes to ServiceControl");
 
-            context.Container.ConfigureComponent(b => new CaptureSagaStateBehavior(context.Settings.EndpointName(), BuildSerializer(context.Settings), b.Build<ServiceControlBackend>()), DependencyLifecycle.SingleInstance);
-
             context.RegisterStartupTask(b => new SagaAuditStartupTask(b.Build<ServiceControlBackend>()));
-        }
-
-        static CaptureSagaStateSerializer BuildSerializer(ReadOnlySettings settings)
-        {
-            var definition = new JsonSerializer();
-
-            var factory = definition.Configure(settings);
-
-            var serializer = factory(new MessageMapper());
-
-            return new CaptureSagaStateSerializer(serializer);
         }
 
         class SagaAuditStartupTask : FeatureStartupTask
