+++ src/ServiceControl.Plugin.Nsb6.SagaAudit/ServiceControl.Plugin.Nsb6.SagaAudit.csproj
@@ -63,7 +63,7 @@
       <ItemGuid>be7306a9-93a4-490d-ac1a-8377f589cebe</ItemGuid>
     </Compile>
     <Compile Include="AssemblyInfo.cs" />
-    <Compile Include="CaptureSagaStateSerializer.cs" />
+    <Compile Include="SagaAuditSerializer.cs" />
     <Compile Include="Messages\SagaChangeInitiator.cs" />
     <Compile Include="Messages\SagaChangeOutput.cs" />
     <Compile Include="Messages\SagaUpdatedMessage.cs" />
